requirements.txt
asgiref==3.7.2
Django==5.0.2
django-restframework==0.0.1
djangorestframework==3.14.0
numpy==1.26.4
opencv-python==4.9.0.80
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
